requirements.txt
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.1.0

